fit card contents dynamically
images and card titles vary in size. I adjusted the images to fit and hide their overflow depending on different screen sizes. i adjusted the card heights to always match the height of the tallest card in the row, and i moved the learn more button to a different section so that they'd always be aligned no matter how long the card title